+++ src/main/java/fr/estecka/variantscit/VCitRegistries.java
@@ -24,9 +24,10 @@ public final class VCitRegistries
 		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("lowercase"),          String::toLowerCase);
 		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("discard_path"),       IStringTransform::DiscardPath);
 		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("discard_namespace"),  IStringTransform::DiscardNamespace);
-		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("sanitize"),           IStringTransform.Sanitize("[^a-zA-Z0-9_.-/:]"));
-		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("sanitize_path"),      IStringTransform.Sanitize("[^a-zA-Z0-9_.-/]"));
-		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("sanitize_namespace"), IStringTransform.Sanitize("[^a-zA-Z0-9_.-]"));
+		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("sanitize"),           IStringTransform.SANITIZE);
+		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("sanitize_path"),      IStringTransform.SANITIZE_PATH);
+		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("sanitize_namespace"), IStringTransform.SANITIZE_NAMESPACE);
+		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("sanitize_auto"),      IStringTransform::AutoSanitize);
 		TRANSFORMS.RegisterMap(Identifier.ofVanilla("regex"),               RegexTransform.MAPCODEC);
 
 		NBT_INPUTS.RegisterUnit(Identifier.ofVanilla("auto"),            INbtInput.AUTO);