+++ Changelog.md
@@ -92,3 +92,4 @@
 - `nbtPath` is now optional and defaults to using the whole component as data.
 - Added simplified mono-string syntax for ItemComponent properties.
 - Format variables now accept upper-cases, numbers and low-dashes
+- Marked module types `custom_data`, `entity_data`, `bucket_entity_data` or `block_entity_data` as deprecated. (Use `component_data` instead.)